Move LiftCoverageCondition to Optics.Internal.Magic
